- Добавил MustBeValueObject (два overload-а: Error и Errors) и WithE…

+++ DirectoryService/src/DirectoryService.Application/Departments/RemoveDepartment/RemoveDepartmentHandler.cs
@@ -1,5 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 using DirectoryService.Application.Abstraction;
+using DirectoryService.Application.Validations;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
 using Shared;
 
@@ -9,17 +11,27 @@ public class RemoveDepartmentHandler : ICommandHandler<RemoveDepartmentCommand, 
 {
     private readonly IDepartmentRepository _departmentRepository;
     private readonly ILogger<RemoveDepartmentHandler> _logger;
+    private readonly IValidator<RemoveDepartmentCommand> _validator;
 
-    public RemoveDepartmentHandler(IDepartmentRepository departmentRepository, ILogger<RemoveDepartmentHandler> logger)
+    public RemoveDepartmentHandler(
+        IDepartmentRepository departmentRepository,
+        ILogger<RemoveDepartmentHandler> logger,
+        IValidator<RemoveDepartmentCommand> validator)
     {
         _departmentRepository = departmentRepository;
         _logger = logger;
+        _validator = validator;
     }
 
     public async Task<Result<Unit, Errors>> Handle(
         RemoveDepartmentCommand command,
         CancellationToken cancellationToken)
     {
+        var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+        if (!validationResult.IsValid)
+        {
+            return validationResult.ToError();
+        }
         _logger.LogInformation("Обработка RemoveDepartmentCommand id:{departmentId}", command.departmentId);
 
         var removeResult = await _departmentRepository.DeleteByIdAsync(command.departmentId, cancellationToken);
